[MIG]Migrated l10n_fr_pos_cert_base to v12.0
