+++ l10n_fr_pos_cert_base/README.rst
@@ -1,0 +1,39 @@
+
+Admin Menu
+==========
+
+Allow to access admin menu:
+---------------------------
+Menu: Setting > User & company > Users
+
+.. image:: /l10n_fr_pos_cert_base/static/description/user_access.png
+   :width: 1100px
+
+|
+|
+|
+
+Allow to configure admin Users on staging
+-----------------------------------------
+Menu: Admin > Admin Users on staging
+
+.. image:: /l10n_fr_pos_cert_base/static/description/user_staging.png
+   :width: 1100px
+
+|
+|
+|
+
+Credits
+=======
+
+Authors
+
+* Druidoo (<https://www.druidoo.io>)
+
+Based on https://github.com/ingadhoc/sale/blob/12.0/pos_voucher/
+
+
+Maintainers
+
+This module is maintained by Druidoo.
