+++ l10n_fr_pos_cert_base/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Copyright (C) 2016-Today: La Louve (<https://cooplalouve.fr>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
