+++ l10n_fr_pos_cert_base/data/res_group_data.xml
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data>
+<odoo>
     <record id="group_funtional_admin" model="res.groups">
         <field name="name">Functional Admin</field>
         <field name="category_id" ref="base.module_category_administration"/>
-        <field name="implied_ids" eval="[(4, ref('base.group_configuration'))]"/>
+        <field name="implied_ids" eval="[(4, ref('base.group_system'))]"/>
     </record>
 
-</data>
-</openerp>
+</odoo>
