+++ l10n_fr_pos_cert_base/models/__init__.py
@@ -1,3 +1,2 @@
-# -*- coding: utf-8 -*-
-from . import base
+from . import res_company
 from . import user_config_settings
