--- l10n_fr_pos_cert_base/models/base/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (C) 2016-Today: La Louve (<https://cooplalouve.fr>)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-
-from . import res_company
