--- l10n_fr_pos_cert_base/models/base/res_company.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-from openerp import api, models, fields
-from openerp import SUPERUSER_ID
-
-
-class ResCompany(models.Model):
-    _inherit = "res.company"
-
-    configuration_user_ids = fields.Many2many(
-        comodel_name='res.users'
-    )
