+++ l10n_fr_pos_cert_base/models/res_company.py
@@ -1,0 +1,6 @@
+from odoo import models, fields
+
+class ResCompany(models.Model):
+    _inherit = "res.company"
+
+    configuration_user_ids = fields.Many2many(comodel_name="res.users")
