--- l10n_fr_pos_cert_base/static/src/js/auto_select.js
@@ -1,5 +1,0 @@
-$(document).ready(function(){
-	$(document).on('focus', '.auto_select input', function(e){
-		$(this).select();
-	})
-});
