--- l10n_fr_pos_cert_base/views/templates.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <template id="assets_backend" name="l10n_fr_pos_cert_base" inherit_id="web.assets_backend">
-            <xpath expr="." position="inside">
-                <link rel="stylesheet" href="/l10n_fr_pos_cert_base/static/src/less/togglebutton.less" type="text/less" t-ignore="true"/>
-                <script type="text/javascript" src="/l10n_fr_pos_cert_base/static/src/js/auto_select.js" />
-                <script type="text/javascript" src="/l10n_fr_pos_cert_base/static/src/js/form_widget.js" />
-            </xpath>
-        </template>
-    </data>
-</openerp>
