+++ l10n_fr_pos_cert_base/views/view_user_config_settings.xml
@@ -1,35 +1,35 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <odoo>
-    <data>
-        <record id="view_user_config_settings" model="ir.ui.view">
-            <field name="name">user.setting.form.view</field>
-            <field name="model">user.config.settings</field>
-            <field name="arch" type="xml">
-                <form string="User Settings">
-                    <header>
-                        <button string="Apply" type="object" name="execute"
-                                class="oe_highlight"/>
-                        <button string="Cancel" type="object" name="cancel"
-                                class="oe_link"/>
-                    </header>
-                    <sheet>
-                        <notebook>
-                            <page string="Admin Users on Staging">
-                                <field name="configuration_user_ids"
-                                       context="{'search_default_no_share':1}"/>
-                            </page>
-                        </notebook>
-                    </sheet>
-                </form>
-            </field>
-        </record>
 
-        <record id="action_user_config" model="ir.actions.act_window">
-            <field name="name">Configure Users</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">user.config.settings</field>
-            <field name="view_mode">form</field>
-            <field name="target">inline</field>
-        </record>
-    </data>
+    <record id="view_user_config_settings" model="ir.ui.view">
+        <field name="name">user.setting.form.view</field>
+        <field name="model">user.config.settings</field>
+        <field name="arch" type="xml">
+            <form string="User Settings">
+                <header>
+                    <button string="Apply" type="object" name="execute"
+                            class="oe_highlight"/>
+                    <button string="Cancel" type="object" name="cancel"
+                            class="oe_link"/>
+                </header>
+                <sheet>
+                    <notebook>
+                        <page string="Admin Users on Staging">
+                            <field name="configuration_user_ids"
+                                   context="{'search_default_no_share':1}"/>
+                        </page>
+                    </notebook>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+    <record id="action_user_config" model="ir.actions.act_window">
+        <field name="name">Configure Users</field>
+        <field name="type">ir.actions.act_window</field>
+        <field name="res_model">user.config.settings</field>
+        <field name="view_mode">form</field>
+        <field name="target">inline</field>
+    </record>
+
 </odoo>
